Add files via upload
This movie recommender system is a user-friendly and interactive application that leverages the power of Streamlit, it is designed to provide personalized movie suggestions to users based on their preferences, viewing history, and behavior. This system employs various machine learning and data analysis techniques to analyze and understand user preferences and movie characteristics .By leveraging this information, it recommends movies that users are likely to enjoy, enhancing their overall movie-watching experience.